Split rest-api part 3
