+++ src/components/dialogs/network-modifications/delete-attaching-line/delete-attaching-line-dialog.js
@@ -18,12 +18,13 @@ import {
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import { deleteAttachingLine, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import { sanitizeString } from '../../dialogUtils';
 import yup from 'components/utils/yup-config';
 import ModificationDialog from '../../commons/modificationDialog';
 import DeleteAttachingLineForm from './delete-attaching-line-form';
 import { useOpenShortWaitFetching } from '../../commons/handle-modification-form';
+import { deleteAttachingLine } from '../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [ATTACHED_LINE_ID]: null,
