+++ src/components/dialogs/network-modifications/generation-dispatch/generation-dispatch-dialog.js
@@ -22,12 +22,13 @@ import {
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import { FetchStatus, generationDispatch } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import yup from 'components/utils/yup-config';
 import { useOpenShortWaitFetching } from '../../commons/handle-modification-form';
 import ModificationDialog from '../../commons/modificationDialog';
 import GenerationDispatchForm from './generation-dispatch-form';
 import { addSelectedFieldToRows } from 'components/utils/dnd-table/dnd-table';
+import { generationDispatch } from '../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [LOSS_COEFFICIENT]: null,
