+++ src/components/dialogs/network-modifications/generator-scaling/generator-scaling-dialog.js
@@ -14,13 +14,13 @@ import { useCallback, useEffect } from 'react';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { VARIATION_TYPE, VARIATIONS } from 'components/utils/field-constants';
 import { getVariationsSchema } from './variation/variation-utils';
-import { generatorScaling } from 'utils/rest-api';
 import {
     VARIATION_TYPES,
     FORM_LOADING_DELAY,
 } from 'components/network/constants';
 import { useOpenShortWaitFetching } from '../../commons/handle-modification-form';
 import { FetchStatus } from 'utils/rest-api';
+import { generatorScaling } from '../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [VARIATION_TYPE]: VARIATION_TYPES.DELTA_P.id,
