+++ src/components/dialogs/network-modifications/line-split-with-voltage-level/line-split-with-voltage-level-dialog.js
@@ -22,7 +22,7 @@ import {
 import PropTypes from 'prop-types';
 import React, { useCallback, useEffect, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
-import { divideLine, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import { sanitizeString } from '../../dialogUtils';
 import yup from 'components/utils/yup-config';
 import ModificationDialog from '../../commons/modificationDialog';
@@ -41,6 +41,7 @@ import { MODIFICATION_TYPES } from 'components/utils/modification-type';
 import { buildNewBusbarSections } from 'components/utils/utils';
 import { useOpenShortWaitFetching } from 'components/dialogs/commons/handle-modification-form';
 import { FORM_LOADING_DELAY } from 'components/network/constants';
+import { divideLine } from '../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [LINE1_ID]: '',
@@ -161,8 +162,6 @@ const LineSplitWithVoltageLevelDialog = ({
 
     const onVoltageLevelCreationDo = useCallback(
         ({
-            studyUuid,
-            currentNodeUuid,
             voltageLevelId,
             voltageLevelName,
             substationId,
