+++ src/components/dialogs/network-modifications/load-scaling/load-scaling-dialog.js
@@ -14,12 +14,13 @@ import { useCallback, useEffect } from 'react';
 import { useSnackMessage } from '@gridsuite/commons-ui';
 import { VARIATION_TYPE, VARIATIONS } from 'components/utils/field-constants';
 import { getVariationsSchema } from './variation/variation-utils';
-import { loadScaling, FetchStatus } from 'utils/rest-api';
+import { FetchStatus } from 'utils/rest-api';
 import {
     FORM_LOADING_DELAY,
     VARIATION_TYPES,
 } from 'components/network/constants';
 import { useOpenShortWaitFetching } from 'components/dialogs/commons/handle-modification-form';
+import { loadScaling } from '../../../../services/study/network-modifications';
 
 const emptyFormData = {
     [VARIATION_TYPE]: VARIATION_TYPES.DELTA_P.id,
