+++ src/components/shortcircuit-analysis-result-tab.js
@@ -6,9 +6,9 @@
  */
 
 import { useNodeData } from './study-container';
-import { fetchShortCircuitAnalysisResult } from '../utils/rest-api';
 import WaitingLoader from './utils/waiting-loader';
 import ShortCircuitAnalysisResult from './shortcircuit-analysis-result';
+import { fetchShortCircuitAnalysisResult } from '../services/study/short-circuit-analysis';
 
 const shortCircuitAnalysisResultInvalidations = ['shortCircuitAnalysisResult'];
