+++ src/services/study/short-circuit-analysis.js
@@ -1,0 +1,51 @@
+import {
+    backendFetch,
+    backendFetchJson,
+    backendFetchText,
+} from '../../utils/rest-api';
+import { getStudyUrlWithNodeUuid } from './index';
+
+export function startShortCircuitAnalysis(studyUuid, currentNodeUuid) {
+    console.info(
+        `Running short circuit analysis on '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+
+    const startShortCircuitAnanysisUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/shortcircuit/run';
+    console.debug(startShortCircuitAnanysisUrl);
+    return backendFetch(startShortCircuitAnanysisUrl, { method: 'put' });
+}
+
+export function stopShortCircuitAnalysis(studyUuid, currentNodeUuid) {
+    console.info(
+        `Stopping short circuit analysis on '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    const stopShortCircuitAnalysisUrl =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/shortcircuit/stop';
+    console.debug(stopShortCircuitAnalysisUrl);
+    return backendFetch(stopShortCircuitAnalysisUrl, { method: 'put' });
+}
+
+export function fetchShortCircuitAnalysisStatus(studyUuid, currentNodeUuid) {
+    console.info(
+        `Fetching short circuit analysis status on '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    const url =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/shortcircuit/status';
+    console.debug(url);
+    return backendFetchText(url);
+}
+
+export function fetchShortCircuitAnalysisResult(studyUuid, currentNodeUuid) {
+    console.info(
+        `Fetching short circuit analysis result on '${studyUuid}' and node '${currentNodeUuid}' ...`
+    );
+    const url =
+        getStudyUrlWithNodeUuid(studyUuid, currentNodeUuid) +
+        '/shortcircuit/result';
+    console.debug(url);
+    return backendFetchJson(url);
+}
